修改详见gitee

+++ DrissionPage/drission.py
@@ -23,8 +23,8 @@ class Drission(object):
     """Drission类用于管理WebDriver对象和Session对象，是驱动器的角色"""
 
     def __init__(self,
-                 driver_or_options: Union[WebDriver, dict, Options, DriverOptions] = None,
-                 session_or_options: Union[Session, dict, SessionOptions] = None,
+                 driver_or_options: Union[WebDriver, dict, Options, DriverOptions, bool] = None,
+                 session_or_options: Union[Session, dict, SessionOptions, bool] = None,
                  ini_path: str = None,
                  proxy: dict = None):
         """初始化，可接收现成的WebDriver和Session对象，或接收它们的配置信息生成对象       \n
@@ -67,6 +67,13 @@ class Drission(object):
             else:
                 self._driver_options = _chrome_options_to_dict(driver_or_options)
 
+    def __del__(self):
+        """关闭对象时关闭浏览器和Session"""
+        try:
+            self.close()
+        except ImportError:
+            pass
+
     @property
     def session(self) -> Session:
         """返回Session对象，如未初始化则按配置信息创建"""
@@ -85,7 +92,7 @@ class Drission(object):
         """
         if self._driver is None:
             if not isinstance(self._driver_options, dict):
-                raise TypeError('Driver options invalid')
+                raise TypeError('无效的Driver配置。')
 
             options = _dict_to_chrome_options(self._driver_options)
 
@@ -98,7 +105,7 @@ class Drission(object):
             # -----------若指定debug端口且该端口未在使用中，则先启动浏览器进程-----------
             if options.debugger_address and _check_port(options.debugger_address) is False:
                 from subprocess import Popen
-                port = options.debugger_address[options.debugger_address.rfind(':') + 1:]
+                port = options.debugger_address.split(':')[-1]
 
                 # 启动浏览器进程，同时返回该进程使用的 chrome.exe 路径
                 chrome_path, self._debugger = _create_chrome(chrome_path, port,
@@ -109,8 +116,8 @@ class Drission(object):
 
             # 反反爬设置
             try:
-                self._driver.execute_script('Object.defineProperty(navigator,"webdriver",{get:() => Chrome,});')
-            except:
+                self._driver.execute_script('Object.defineProperty(navigator,"webdriver",{get:() => undefined,});')
+            except Exception:
                 pass
 
             # self._driver.execute_cdp_cmd(
@@ -176,29 +183,60 @@ class Drission(object):
             self.debugger_progress.kill()
             return
 
-        address = self.driver_options.get('debugger_address', '').split(':')
-        if len(address) == 1:
+        pid = self.get_browser_progress_id()
+        from os import popen
+        if pid and popen(f'tasklist | findstr {pid}').read().lower().startswith('chrome.exe'):
+            print('kill')
+            popen(f'taskkill /pid {pid} /F')
+        else:
             self.close_driver()
 
-        elif len(address) == 2:
+    def get_browser_progress_id(self) -> Union[str, None]:
+        """获取浏览器进程id"""
+        if self.debugger_progress:
+            return self.debugger_progress.pid
+
+        address = self.driver_options.get('debugger_address', '').split(':')
+        if len(address) == 2:
             ip, port = address
             if ip not in ('127.0.0.1', 'localhost') or not port.isdigit():
-                return
+                return None
 
             from os import popen
-            progresses = popen(f'netstat -nao | findstr :{port}').read().split('\n')
             txt = ''
+            progresses = popen(f'netstat -nao | findstr :{port}').read().split('\n')
             for progress in progresses:
                 if 'LISTENING' in progress:
                     txt = progress
                     break
-
             if not txt:
-                return
+                return None
 
-            pid = txt[txt.rfind(' ') + 1:]
-            if popen(f'tasklist | findstr {pid}').read().lower().startswith('chrome.exe'):
-                popen(f'taskkill /pid {pid} /F')
+            return txt.split(' ')[-1]
+
+    def hide_browser(self) -> None:
+        """隐藏浏览器界面"""
+        self._show_or_hide_browser()
+
+    def show_browser(self) -> None:
+        """显示浏览器界面"""
+        self._show_or_hide_browser(False)
+
+    def _show_or_hide_browser(self, hide: bool = True) -> None:
+        try:
+            from win32gui import ShowWindow
+            from win32con import SW_HIDE, SW_SHOW
+        except ImportError:
+            raise ImportError('请先安装：pip install pypiwin32')
+
+        pid = self.get_browser_progress_id()
+        if not pid:
+            print('只有设置了debugger_address参数才能使用 show_browser() 和 hide_browser()')
+            return
+        hds = _get_chrome_hwnds_from_pid(pid)
+        sw = SW_HIDE if hide else SW_SHOW
+        for hd in hds:
+            ShowWindow(hd, sw)
 
     def set_cookies(self,
                     cookies: Union[RequestsCookieJar, list, tuple, str, dict],
@@ -241,7 +279,7 @@ class Drission(object):
                         url = extract(browser_domain)
                         cookie_domain = f'{url.domain}.{url.suffix}'
                     else:
-                        raise ValueError('There is no domain name in the cookie or the browser has not visited a URL.')
+                        raise ValueError('cookie中没有域名或浏览器未访问过URL。')
 
                     cookie['domain'] = cookie_domain
 
@@ -342,13 +380,6 @@ class Drission(object):
         if self._session:
             self.close_session()
 
-    def __del__(self):
-        """关闭对象时关闭浏览器和Session"""
-        try:
-            self.close()
-        except ImportError:
-            pass
-
 
 def _check_port(debugger_address: str) -> Union[bool, None]:
     """检查端口是否被占用                               \n
@@ -438,7 +469,7 @@ def _create_driver(chrome_path: str, driver_path: str, options: Options) -> WebD
         if driver_path:
             try:
                 return webdriver.Chrome(driver_path, options=options)
-            except:
+            except Exception:
                 pass
 
         # 当找不到 driver 且 chrome_path 为 None 时，说明安装的版本过高，改在系统路径中查找
@@ -451,8 +482,28 @@ def _create_driver(chrome_path: str, driver_path: str, options: Options) -> WebD
                 options.binary_location = chrome_path
                 try:
                     return webdriver.Chrome(driver_path, options=options)
-                except:
+                except Exception:
                     pass
 
-    print('无法启动，请检查chromedriver版本与Chrome是否匹配，并手动设置。')
+    print('无法启动，请检查浏览器路径，或手动设置chromedriver。\n下载地址：http://npm.taobao.org/mirrors/chromedriver/')
     exit(0)
+
+
+def _get_chrome_hwnds_from_pid(pid) -> list:
+    # 通过PID查询句柄ID
+    try:
+        from win32gui import IsWindow, GetWindowText, EnumWindows
+        from win32process import GetWindowThreadProcessId
+    except ImportError:
+        raise ImportError('请先安装win32gui，pip install pypiwin32')
+
+    def callback(hwnd, hds):
+        if IsWindow(hwnd) and '- Google Chrome' in GetWindowText(hwnd):
+            _, found_pid = GetWindowThreadProcessId(hwnd)
+            if str(found_pid) == str(pid):
+                hds.append(hwnd)
+            return True
+
+    hwnds = []
+    EnumWindows(callback, hwnds)
+    return hwnds
